layout/generic/test/test_bug449653.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=449653
-->
<head>
  <title>Test for Bug 449653</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style>
    iframe {
      width: 500px;
      height: 300px;
    }
  </style>
</head>
<body onload="doTest()">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=449653">Mozilla Bug 449653</a>
<p id="display"></p>
<div id="content">
<iframe src="file_bug449653_1.html" id="f1"></iframe>
<iframe id="f2" src="file_bug449653_1_ref.html"></iframe>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 449653 **/
SimpleTest.waitForExplicitFinish();

var f1 = document.getElementById("f1");
var f2 = document.getElementById("f2");

function doTest() {
  ok(compareSnapshots(snapshotWindow(f1.contentWindow),
                      snapshotWindow(f2.contentWindow), true)[0],
     "No red should be shown");

  SimpleTest.finish();
}
</script>
</pre>
</body>
</html>
